move middleware to separate module
